feature(collect_info): implement getting column index by cell value
Fixes https://github.com/igorsimb/atani-report/issues/1
`conclusions` function iterates over a hard coded column index, but the actual values can be in different columns.
This change creates a function that looks for a specific value in a cell and returns its column index. Now the conclusions can be stored in any column.
Also, changed a variable name for conclusions sheet for clarity